[STDINT] Applied overall variable with uintxx_t

+++ include/arch/armv7/generic_timer.h
@@ -1,12 +1,14 @@
 #ifndef __GENERIC_TIMER_H___
 #define __GENERIC_TIMER_H___
 
+#include <stdint.h>
+
 #ifdef CONFIG_C99
 #include "c99.h"
 #endif
 
 /* Generic Timer */
-#define read_cntfrq()           ({ unsigned int rval; asm volatile(\
+#define read_cntfrq()           ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c0, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -14,7 +16,7 @@
                                 " mcr     p15, 0, %0, c14, c0, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cnthctl()          ({ unsigned int rval; asm volatile(\
+#define read_cnthctl()          ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 4, %0, c14, c1, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -22,7 +24,7 @@
                                 " mcr     p15, 4, %0, c14, c1, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cnthp_ctl()        ({ unsigned int rval; asm volatile(\
+#define read_cnthp_ctl()        ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 4, %0, c14, c2, 1\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -30,17 +32,17 @@
                                 " mcr     p15, 4, %0, c14, c2, 1\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cnthp_cval()       ({ unsigned int v1, v2; asm volatile(\
+#define read_cnthp_cval()       ({ uint32_t v1, v2; asm volatile(\
                                 " mrrc     p15, 6, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
+                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
 
 #define write_cnthp_cval(val) 	asm volatile(\
                               	  " mcrr     p15, 6, %0, %1, c14\n\t" \
 								  : : "r" ((val) & 0xFFFFFFFF), "r" ((val) >> 32) \
 								  : "memory", "cc")
 
-#define read_cnthp_tval()       ({ unsigned int rval; asm volatile(\
+#define read_cnthp_tval()       ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 4, %0, c14, c2, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -48,7 +50,7 @@
                                 " mcr     p15, 4, %0, c14, c2, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntkctl()          ({ unsigned int rval; asm volatile(\
+#define read_cntkctl()          ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c1, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -56,7 +58,7 @@
                                 " mcr     p15, 0, %0, c14, c1, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntp_ctl()         ({ unsigned int rval; asm volatile(\
+#define read_cntp_ctl()         ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c2, 1\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -64,17 +66,17 @@
                                 " mcr     p15, 0, %0, c14, c2, 1\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntp_cval()        ({ unsigned int v1, v2; asm volatile(\
+#define read_cntp_cval()        ({ uint32_t v1, v2; asm volatile(\
                                 " mrrc     p15, 2, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
+                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
 
 #define write_cntp_cval(val)  	asm volatile(\
                               	  " mcrr     p15, 2, %0, %1, c14\n\t" \
 								  : : "r" ((val) & 0xFFFFFFFF), "r" ((val) >> 32) \
 								  : "memory", "cc")
 
-#define read_cntp_tval()        ({ unsigned int rval; asm volatile(\
+#define read_cntp_tval()        ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c2, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -82,12 +84,12 @@
                                 " mcr     p15, 0, %0, c14, c2, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntpct()           ({ unsigned int v1, v2; asm volatile(\
+#define read_cntpct()           ({ uint32_t v1, v2; asm volatile(\
                                 " mrrc     p15, 0, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
+                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
 
-#define read_cntv_ctl()         ({ unsigned int rval; asm volatile(\
+#define read_cntv_ctl()         ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c3, 1\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -95,17 +97,17 @@
                                 " mcr     p15, 0, %0, c14, c3, 1\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntv_cval()        ({ unsigned int v1, v2; asm volatile(\
+#define read_cntv_cval()        ({ uint32_t v1, v2; asm volatile(\
                                 " mrrc     p15, 3, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
+                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
 
 #define write_cntv_cval(val)  asm volatile(\
                               " mcrr     p15, 3, %0, %1, c14\n\t" \
                               : : "r" ((val) & 0xFFFFFFFF), "r" ((val) >> 32) \
                               : "memory", "cc")
 
-#define read_cntv_tval()        ({ unsigned int rval; asm volatile(\
+#define read_cntv_tval()        ({ uint32_t rval; asm volatile(\
                                 " mrc     p15, 0, %0, c14, c3, 0\n\t" \
                                 : "=r" (rval) : : "memory", "cc"); rval; })
 
@@ -113,15 +115,15 @@
                                 " mcr     p15, 0, %0, c14, c3, 0\n\t" \
                                 : : "r" ((val)) : "memory", "cc")
 
-#define read_cntvct()           ({ unsigned int v1, v2; asm volatile(\
+#define read_cntvct()           ({ uint32_t v1, v2; asm volatile(\
                                 " mrrc     p15, 1, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
+                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
 
-#define read_cntvoff()          ({ unsigned int v1, v2; asm volatile(\
+#define read_cntvoff()          ({ uint32_t v1, v2; asm volatile(\
                                 " mrrc     p15, 4, %0, %1, c14\n\t" \
                                 : "=r" (v1), "=r" (v2) : : "memory", "cc"); \
-                                (((unsigned long long)v2 << 32) + (unsigned long long)v1); })
+                                (((uint64_t)v2 << 32) + (uint64_t)v1); })
 
 #define write_cntvoff(val)    asm volatile(\
                               " mcrr     p15, 4, %0, %1, c14\n\t" \
